Change to go.mod mode and upgrade the redis dependency package，Improve the code
