+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/heiyeluren/koala
+
+go 1.16
+
+require github.com/gomodule/redigo v1.8.5
